Add module for teachers and teacher class
