+++ teacher.rb
@@ -1,0 +1,12 @@
+class Teacher
+
+  attr_reader :age, :salary, :phase, :target_raise
+  attr_accessor :name
+
+  def initialize(options={})
+    @age = options.fetch(:age, 0)
+    @name = options.fetch(:name, "")
+    @phase = 3
+  end
+
+end
